+++ src/V3/Models/Import/FailureCollectionInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Dotdigital\V3\Models\Import;
+
+interface FailureCollectionInterface
+{
+}