+++ tests/V3/Unit/ContactImportTest.php
@@ -3,12 +3,12 @@
 namespace V3\Unit;
 
 use Dotdigital\V3\Models\Contact;
-use Dotdigital\V3\Models\Contact\Identifiers;
 use Dotdigital\V3\Models\Contact\Import;
+use Dotdigital\V3\Models\Contact\Identifiers;
 use Dotdigital\V3\Models\Contact\Import\Failure;
-use Dotdigital\V3\Models\Contact\Import\FailureDetail;
-use Dotdigital\V3\Models\Contact\Import\FailureDetailCollection;
 use Dotdigital\V3\Models\Contact\Import\Summary;
+use Dotdigital\V3\Models\Contact\Import\Failure\FailureDetail;
+use Dotdigital\V3\Models\Contact\Import\Failure\FailureDetailCollection;
 use Dotdigital\V3\Models\ContactCollection;
 use PHPUnit\Framework\TestCase;